Allow body validator to be created where possible
The body validator was not getting created because the reference schema
does not contain the requestPath which in turn is because the way
external file is split (it moves all path of user to a separate file).

As referencedBodySchema is not needed to create the validator that is
needed I have relaxed the guard statement in handleBodyValidation().